+++ src/main/kotlin/de/fuballer/mcendgame/main/component/item/custom/crystal/CrystalItems.kt
@@ -1,7 +1,7 @@
 package de.fuballer.mcendgame.main.component.item.custom.crystal
 
 import de.fuballer.mcendgame.main.component.item.custom.crystal.item.CalibrationCrystalItem
-import de.fuballer.mcendgame.main.component.item.custom.crystal.item.corruption.CorruptionCrystalItem
+import de.fuballer.mcendgame.main.component.item.custom.crystal.item.CorruptionCrystalItem
 import de.fuballer.mcendgame.main.component.item.custom.crystal.item.PermutationCrystalItem
 import de.fuballer.mcendgame.main.component.item.custom.crystal.item.ReforgeCrystalItem
 import de.fuballer.mcendgame.main.component.item.custom.crystal.item.SacrificeCrystalItem
